requirements.txt
python-terraform==0.10.1
PyYAML==6.0